v4.7.2556.1 - BugFixes + VERSIONING change
Versioning Changes:
      - date sig now BEFORE build (i.e., MAJ.MIN.SIG.BUILD).
      - date sig now keyed to 2011-06-15 (Wednesday) - date of first 'script' created for what would become S4.

Bug fix for load_header_details (bigint for backup size) + ripple of changes up into restore_databases.

+++ S4 Restore/Utilities/load_header_details.sql
@@ -30,7 +30,7 @@ GO
 CREATE PROC dbo.load_header_details 
 	@BackupPath			nvarchar(800), 
 	@BackupDate			datetime		OUTPUT, 
-	@BackupSize			int				OUTPUT, 
+	@BackupSize			bigint			OUTPUT, 
 	@Compressed			bit				OUTPUT, 
 	@Encrypted			bit				OUTPUT
 
@@ -121,7 +121,7 @@ AS
 	-- Return Output Details: 
 	SELECT 
 		@BackupDate = [BackupFinishDate], 
-		@BackupSize = CAST((ISNULL([CompressedBackupSize], [BackupSize])) AS int), 
+		@BackupSize = CAST((ISNULL([CompressedBackupSize], [BackupSize])) AS bigint), 
 		@Compressed = [Compressed], 
 		@Encrypted = CASE WHEN EncryptorThumbprint IS NOT NULL THEN 1 ELSE 0 END
 	FROM
